+++ saleor/account/tests/test_account.py
@@ -172,6 +172,8 @@ def test_address_as_data(address):
         "country": "PL",
         "country_area": "",
         "phone": "+48713988102",
+        "metadata": {},
+        "private_metadata": {},
     }