+++ saleor/payment/gateways/adyen/plugin.py
@@ -650,6 +650,7 @@ class AdyenGatewayPlugin(BasePlugin):
             error=None,
             raw_response={},
             transaction_already_processed=bool(transaction_already_processed),
+            psp_reference=token,
         )
 
     def refund_payment(