+++ saleor/graphql/core/descriptions.py
@@ -21,6 +21,7 @@ ADDED_IN_37 = "\n\nAdded in Saleor 3.7."
 ADDED_IN_38 = "\n\nAdded in Saleor 3.8."
 ADDED_IN_39 = "\n\nAdded in Saleor 3.9."
 ADDED_IN_310 = "\n\nAdded in Saleor 3.10."
+ADDED_IN_311 = "\n\nAdded in Saleor 3.11."
 
 
 PREVIEW_FEATURE = (